requirements.txt
alembic==0.7.4
amqp==1.4.6
anyjson==0.3.3
awesome-slugify==1.6.2
billiard==3.3.0.19
celery==3.1.17
decorator==3.4.0
enum==0.4.4
Flask==0.10.1
Flask-Migrate==1.3.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
humanize==0.5.1
itsdangerous==0.24
Jinja2==2.7.3
kombu==3.0.24
Mako==1.0.1
MarkupSafe==0.23
mysql-connector-python==2.0.2
nose==1.3.4
pbr==0.10.7
regex==2014.12.24
SQLAlchemy==0.9.8
sqlparse==0.1.14
subprocess32==3.2.6
Tempita==0.5.2
Unidecode==0.4.17
Werkzeug==0.10.1
